refactor(tools): split source meta builders
